Stelt de finale diameter-hoogtecurves grafisch voor in een rapport

De functie genereert een rapport (.html-bestand) in de working directory (of opgegeven directory) met informatie en grafieken van de finale modellen.

Deze functie rendert het bestand dhcurverapport.Rmd, dat voor elke boomsoort-domeincombinatie child dhcurve.Rmd toevoegt.

dhcurvesrapport(
  OutputIVANHO,
  Datalijst,
  Bestandsnaam = "dhcurves.html",
  KleurUitbreiding = FALSE,
  verbose = TRUE,
  PathWD = getwd()
)

Arguments

OutputIVANHO

Lijst met geschatte hoogte per omtrekklasse (dataframe zoals gegenereerd door de functie outputIVANHO()).

Datalijst

Datasets met gemeten waarden die gebruikt zijn voor de modellen, zoals gegenereerd door de functie initiatie().

Bestandsnaam

Een naam voor het rapport (.html-bestand) dat gegenereerd wordt, bestaande uit een string die eindigt op .html

KleurUitbreiding

Moeten de metingen van de uitbreiding in een andere kleur weergegeven worden? Default is FALSE, waarbij (eventuele) metingen van een uitbreiding niet in een andere kleur weergegeven worden.

verbose

Dit geeft de toestand van het systeem aan en zorgt ervoor dat boodschappen niet onnodig gegeven worden. (Defaultwaarde behouden.)

PathWD

Het path van de working directory, dus het path waarin het validatierapport opgeslagen moet worden (default wordt het in de op dat moment actieve working directory opgeslagen).

Value

De functie genereert in de working directory (of opgegeven directory) een rapport (html) met de te controleren modellen. Hierin wordt per model (boomsoort-domeincombinatie) de volgende algemene informatie vermeld: boomsoort, domein (en ID), modeltype, aantal metingen, RMSE en bruikbaar interval.

Daaronder wordt telkens grafisch volgende info weergegeven:

een puntenwolk die de metingen voorstelt (geen individuele metingen, maar een jitter)

curve van het domeinmodel, aangeduid met zwarte ruiten

grenzen van het bruikbaar interval (curves eindigen bij de klassenmiddens die overeenkomen met deze grenzen)