Fit afgeleid model op basis van het basismodel en de opgegeven dataset

Functie die op basis van de opgegeven dataset een model opstelt per boomsoort-domeincombinatie door verschuiving van het Vlaams model voor die boomsoort (afgeleid uit het basismodel). Dit komt overeen met het fitten van het lineair model \(Hoogte_{afgeleid} \sim \delta A + Hoogte_{Vlaams}\) op basis van de opgegeven dataset, waarbij \(Hoogte_{Vlaams}\) de hoogteschatting is volgens het Vlaams model.

Usage

fit.afgeleid(Data.afgeleid, Basismodel)

Arguments

Data.afgeleid

De dataframe "Afgeleid" uit de list die teruggegeven wordt bij de functie initiatie(). (Een dataframe met minimaal de velden BMS (boomsoort), IDbms (identificatienummer van de boomsoort), DOMEIN_ID (identificatienummer van het domein), BOS_BHI (domeinnaam), HOOGTE, Omtrek, logOmtrek, logOmtrek2, Status (van data: "niet gecontroleerd", "gecontroleerd",...), nBomen (aantal bomen per boomsoort-domeincombinatie), nBomenOmtrek05 (aantal bomen per boomsoort-domeincombinatie met omtrek > 0.5 m), nBomenInterval (aantal bomen per boomsoort-domeincombinatie binnen bruikbaar interval), nBomenIntervalOmtrek05 (aantal bomen per boomsoort-domeincombinatie binnen bruikbaar interval met omtrek > 0.5 m), Q5k en Q95k ("bruikbaar interval").)

Basismodel

Model per boomsoort zoals teruggegeven door de functie fit.basis(): tibble met de velden BMS (boomsoort) en Model (lme-object met het gefit mixed model voor die boomsoort)

Value

List met 2 tibbles:

tibble met velden BMS (boomsoort), DOMEIN_ID en Model (lm-object van model per boomsoort-domeincombinatie)

tibble met de ingevoerde dataset, waaraan het veld H_VL_finaal (de hoogteschatting volgens het Vlaams model) toegevoegd is