Hoogteschatting op basis van opgegeven afgeleid model

Functie die de gemiddelde hoogte per omtrekklasse schat voor de domeincurves en Vlaamse curves van het opgegeven afgeleid model. De teruggegeven dataframe kan gebruikt worden om grafieken te maken of afwijkende metingen te bestuderen. Opgelet! In tegenstelling tot de meeste functies van dit package werkt deze functie op basis van 1 model en de bijhorende meetgegevens. Zie voorbeeld voor een methode om deze functie te kunnen toepassen op de volledige dataset Afgeleidmodel.

Usage

hoogteschatting.afgeleid(Domeinsoortmodel, Domeinsoortdata)

Arguments

Domeinsoortmodel

verschoven Vlaams model voor 1 boomsoort-domeincombinatie

Domeinsoortdata

de gegevens die hierbij horen: meetresultaten voor 1 boomsoort-domeincombinatie

Value

dataframe met de meetresultaten en de schattingen van de hoogtes voor het domeinmodel en de Vlaamse model

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#Datasets inladen en het basismodel en afgeleid model berekenen
Data <- testdataset()
Datalijst <- initiatie(Data)
Data.basis <- Datalijst[["Basis"]]
Basismodel <- fit.basis(Data.basis)
Data.afgeleid <- Datalijst[["Afgeleid"]]
Afgeleidmodel <- fit.afgeleid(Data.afgeleid, Basismodel)

#De hoogteschatting voor een afgeleid model
Afgeleidmodel[[1]] %>%
  inner_join(
    Afgeleidmodel[[2]],
    by = c("BMS", "DOMEIN_ID")
  ) %>%
  group_by(
    .data$BMS,
    .data$DOMEIN_ID
  ) %>%
  do(
    hoogteschatting.afgeleid(.$Model[[1]],
                              select(., -Model))
  ) %>%
  ungroup()
#> # A tibble: 74 × 22
#>    DOMEIN_ID BMS    BOS_BHI nBomenOmtrek05 nBomenInterval nBomenIntervalOmtrek05
#>    <chr>     <chr>  <chr>            <int>          <int>                  <int>
#>  1 G         testb… Domein…             39             50                     39
#>  2 G         testb… Domein…             39             50                     39
#>  3 G         testb… Domein…             39             50                     39
#>  4 G         testb… Domein…             39             50                     39
#>  5 G         testb… Domein…             39             50                     39
#>  6 G         testb… Domein…             39             50                     39
#>  7 G         testb… Domein…             39             50                     39
#>  8 G         testb… Domein…             39             50                     39
#>  9 G         testb… Domein…             39             50                     39
#> 10 G         testb… Domein…             39             50                     39
#> # ℹ 64 more rows
#> # ℹ 16 more variables: nBomen <int>, Q5k <dbl>, Q95k <dbl>, Omtrek <dbl>,
#> #   H_VL_finaal <dbl>, IDbms <dbl>, C13 <dbl>, HOOGTE <dbl>, Status <chr>,
#> #   ID <chr>, Rijnr <int>, logOmtrek <dbl>, logOmtrek2 <dbl>, Q5 <dbl>,
#> #   Q95 <dbl>, H_D_finaal <dbl>