Hoogteschatting op basis van opgegeven basismodel

Functie die de gemiddelde hoogte per omtrekklasse schat voor de domeincurves en Vlaamse curves van het opgegeven basismodel. De teruggegeven dataframe kan gebruikt worden om grafieken te maken of afwijkende metingen te bestuderen. Opgelet! In tegenstelling tot de meeste functies van dit package werkt deze functie op basis van 1 model en de bijhorende meetgegevens. Zie voorbeelden voor een methode om deze functie te kunnen toepassen op het Basismodel of het Lokaalmodel en Data.lokaal.

hoogteschatting.basis(
  Soortmodel,
  Soortdata,
  Typemodel,
  BMS,
  Uitbreiding = FALSE
)

Arguments

Soortmodel

model voor één boomsoort (basis) of één boomsoort-domeincombinatie (lokaal)

Soortdata

meetgegevens van boomsoort (basis) of boomsoort-domeincombinatie (lokaal)

Typemodel

"Basis" of "Lokaal"?

BMS

Boomsoort

Uitbreiding

Gaat het hier over berekening voor een uitbreiding? (Facultatief argument om met hogere hoogteklassen te kunnen omgaan.) Default is FALSE, wat betekent dat het niet over een uitbreiding gaat.

Value

dataframe met de meetresultaten en de schattingen van de hoogtes voor het domeinmodel en de Vlaamse model

Examples

library(dplyr)

#Dataset inladen en het basismodel berekenen
Data <- testdataset()
Datalijst <- initiatie(Data)
Data.basis <- Datalijst[["Basis"]]
Basismodel <- fit.basis(Data.basis)

#De hoogteschatting voor een basismodel
Basismodel %>%
  rowwise() %>%
  do(
    hoogteschatting.basis(.$Model, .$Model$data, "Basis", .$BMS)
  ) %>%
  ungroup()
#> # A tibble: 1,118 × 23
#>    DOMEIN_ID BOS_BHI nBomenOmtrek05 nBomenInterval nBomenIntervalOmtrek05 nBomen
#>    <chr>     <chr>            <int>          <int>                  <int>  <int>
#>  1 A         Domein…            174            200                    174    200
#>  2 B         Domein…             87             99                     87    100
#>  3 C         Domein…            423            500                    423    500
#>  4 D         Domein…             85            100                     85    100
#>  5 E         Domein…             92            100                     92    100
#>  6 F         Domein…             82            100                     82    100
#>  7 A         Domein…            174            200                    174    200
#>  8 A         Domein…            174            200                    174    200
#>  9 A         Domein…            174            200                    174    200
#> 10 A         Domein…            174            200                    174    200
#> # ℹ 1,108 more rows
#> # ℹ 17 more variables: Q5k <dbl>, Q95k <dbl>, Omtrek <dbl>, H_D_finaal <dbl>,
#> #   H_VL_finaal <dbl>, IDbms <dbl>, C13 <dbl>, HOOGTE <dbl>, Status <chr>,
#> #   ID <chr>, Rijnr <int>, logOmtrek <dbl>, logOmtrek2 <dbl>, Q5 <dbl>,
#> #   Q95 <dbl>, nExtra <dbl>, BMS <chr>

#Dataset inladen en het lokaal model berekenen
Data.lokaal <- Data.basis %>%
  filter(DOMEIN_ID == "A")
Lokaalmodel <- fit.lokaal(Data.lokaal)

#De hoogteschatting voor een lokaal model
Lokaalmodel %>%
  inner_join(
    Data.lokaal,
    by = c("BMS", "DOMEIN_ID")
  ) %>%
  group_by(
    BMS,
    DOMEIN_ID
  ) %>%
  do(
    hoogteschatting.basis(.$Model[[1]],
                           select(., -Model),
                           "Lokaal", unique(.$BMS))
  ) %>%
  ungroup()
#> # A tibble: 203 × 23
#>    DOMEIN_ID BOS_BHI nBomenOmtrek05 nBomenInterval nBomenIntervalOmtrek05 nBomen
#>    <chr>     <chr>            <int>          <int>                  <int>  <int>
#>  1 A         Domein…            174            200                    174    200
#>  2 A         Domein…            174            200                    174    200
#>  3 A         Domein…            174            200                    174    200
#>  4 A         Domein…            174            200                    174    200
#>  5 A         Domein…            174            200                    174    200
#>  6 A         Domein…            174            200                    174    200
#>  7 A         Domein…            174            200                    174    200
#>  8 A         Domein…            174            200                    174    200
#>  9 A         Domein…            174            200                    174    200
#> 10 A         Domein…            174            200                    174    200
#> # ℹ 193 more rows
#> # ℹ 17 more variables: Q5k <dbl>, Q95k <dbl>, Omtrek <dbl>, H_D_finaal <dbl>,
#> #   IDbms <dbl>, C13 <dbl>, HOOGTE <dbl>, Status <chr>, ID <chr>, Rijnr <int>,
#> #   logOmtrek <dbl>, logOmtrek2 <dbl>, Q5 <dbl>, Q95 <dbl>, VoorModelFit <lgl>,
#> #   nExtra <dbl>, BMS <chr>