Berekent de geschatte hoogtes per omtrekklasse

De functie outputIVANHO() voert de berekeningen uit die leiden tot de gevraagde tabel om in IVANHO te importeren: een tabel met per boomsoort, domein en omtrekklasse een schatting van de hoogte. Als de curve een maximum hoogte vertoont binnen het bestudeerde interval, wordt deze maximumwaarde als hoogte meegegeven aan alle omtrekklassen hoger dan de omtrekklasse van dit maximum. (Dus verschillend van de validatierapporten daalt de hoogte hier niet terug na het maximum, maar de hoogste waarde wordt aangehouden.) Als de curve een minimum hoogte vertoont binnen het bestudeerde interval, wordt deze minimumwaarde als hoogte meegegeven aan alle omtrekklassen lager dan de omtrekklasse van dit minimum. Als voor een bepaalde omtrekklasse de hoogteschatting lager is dan 2,5 m, dan wordt deze omtrekklasse weggelaten.

Voor deze functie worden volgende hulpfuncties aangeroepen: hoogteschatting.basis(), hoogteschatting.afgeleid() en curvekarakteristieken().

Usage

outputIVANHO(
  Basismodel = NULL,
  Afgeleidmodel = NULL,
  Lokaalmodel = NULL,
  Data.lokaal = NULL,
  Uitbreiding = NULL
)

Arguments

Basismodel

Model per boomsoort zoals teruggegeven door de functie fit.basis(): tibble met de velden BMS (boomsoort) en Model (lme-object met het gefit mixed model voor die boomsoort).

Afgeleidmodel

Model per boomsoort-domeincombinatie zoals teruggegeven door de functie fit.afgeleid(): list met 2 tibbles.

Lokaalmodel

Model per boomsoort-domeincombinatie zoals teruggegeven door de functie fit.lokaal(): tibble met de velden BMS (boomsoort), DOMEIN_ID en Model (lm-object met het gefit lineair model voor die boomsoort-domeincombinatie).

Data.lokaal

Dataset op basis waarvan het opgegeven lokaal model berekend is.

Uitbreiding

Dataset met velden BMS, DOMEIN_ID en MaxOmtrek die teruggegeven wordt door functie validatie.uitbreiding(). Als zowel een Basismodel en als Lokaal model opgegeven worden, voeg dan de uitvoer van validatie.uitbreiding() van beide modeltypen samen.

Value

Dataframe met geschatte hoogtes per domein en per boomsoort met velden:

DOMEIN_ID: domeincode

BOS_BHI: domeinnaam

IDbms (identificatienummer van de boomsoort)

BMS: boomsoort

Modeltype ("basismodel"(= eigen model op basis van mixed model) of "afgeleid model"(= verschoven Vlaams model, afgeleid van fixed factor uit basismodel) of "Vlaams model"(= fixed factor uit basismodel, niet toegevoegd omdat niet relevant) of "lokaal model"(= apart model voor 1 boomsoort-domeincombinatie) of "geen model"(= boomsoort-domeincombinatie waarvoor geen model berekend kan worden))

Omtrek: klassenmidden van omtrekklasse

OmtrekklassetypeID: het overeenkomstige identificatienummer van de omtrekklasse

Omtrekklasse

Hoogte: de geschatte hoogte

RMSE (root mean square error, zie vignet voor meer info: vignette("Handleiding", package = "dhcurve"))