Berekent de modelparameters per domein

De functie resultaat berekent op basis van de opgegeven modellen voor elke boomsoort-domeincombinatie de modelparameters A, B en C voor een model van de vorm \(Hoogte \sim A + B\log(Omtrek) + C\log(Omtrek)^2\).

Voor deze functie worden volgende hulpfuncties aangeroepen: modelparameters(), rsme.basis() en rsme.afgeleid()

Verder worden parameters A, B en C uit het model gehaald en een aantal in de functie initiatie() berekende gegevens toegevoegd.

resultaat(
  Basismodel = NULL,
  Afgeleidmodel = NULL,
  Lokaalmodel = NULL,
  Data.lokaal = NULL,
  Data.onbruikbaar = NULL
)

Arguments

Basismodel

Model per boomsoort zoals teruggegeven door de functie fit.basis(): tibble met de velden BMS (boomsoort) en Model (lme-object met het gefit mixed model voor die boomsoort).

Afgeleidmodel

Model per boomsoort-domeincombinatie zoals teruggegeven door de functie fit.afgeleid(): list met 2 tibbles.

Lokaalmodel

Model per boomsoort-domeincombinatie zoals teruggegeven door de functie fit.lokaal(): tibble met de velden BMS (boomsoort), DOMEIN_ID en Model (lm-object met het gefit lineair model voor die boomsoort-domeincombinatie).

Data.lokaal

Dataset op basis waarvan het opgegeven lokaal model berekend is.

Data.onbruikbaar

Evt. lijst met meetresultaten van boomsoort-domeincombinaties waarvoor geen model opgesteld kan worden, in de vorm van een dataframe zoals de dataframe Rest uit de list die door de functie initiatie() teruggegeven wordt.

Value

Dataframe met modellen per domein en per boomsoort met velden:

DomeinID (identificatienummer voor domein)

BMS (boomsoort)

Modeltype ("basismodel"(= eigen model op basis van mixed model) of "afgeleid model"(= verschoven Vlaams model, afgeleid van fixed factor uit basismodel) of "Vlaams model"(= fixed factor uit basismodel, niet toegevoegd omdat niet relevant) of "lokaal model"(= eigen model voor 1 boomsoort-domeincombinatie) of "geen model"(= boomsoort-domeincombinatie waarvoor minstens 1 boom opgemeten is maar geen model berekend kan worden))

parameters A, B en C (zie 'Description' bovenaan)

bruikbaar interval (Q5k en Q95k, zie vignet voor meer info)

RMSE (root mean square error, zie vignet voor meer info)

nBomen (totaal aantal opgemeten bomen met omtrek tussen 0,2 en 3,0 m)

nBomenOmtrek05 (aantal metingen met omtrek tussen 0.5 en 2,8 m, dus waarop afgeleid model gebaseerd is)

nBomenInterval (aantal metingen binnen bruikbaar interval, dus waarop basismodel of lokaal model gebaseerd is)

nBomenIntervalOmtrek05 (aantal metingen binnen bruikbaar interval met omtrek > 0.5 m, dus waarop RMSE-berekening gebaseerd is)

nExtra (aantal metingen boven bruikbaar interval waarop een eventuele uitbreiding gebaseerd is)

evt. kan een tweede dataframe toegevoegd worden met Vlaamse modellen per boomsoort, of deze kan toegevoegd worden aan de vorige dataframe, waarbij DomeinID leeg gelaten wordt of een specifieke waarde "Vlaams model" krijgt