Berekent RMSE van afgeleid model

Deze functie berekent de RMSE op basis van verschil tussen de hoogteschatting volgens het domeinmodel en de gemeten hoogte (voor omtrekklassen > 0.5 m). Dit is slechts een deel van de totale RMSE van het afgeleid model, want behalve deze RMSE van de verschuiving moet hier ook de RMSE van het Vlaams model (basismodel) in rekening gebracht worden. Opgelet! In tegenstelling tot de meeste functies van dit package werkt deze functie op basis van de meetgegevens van 1 model. Zie voorbeeld voor een methode om deze functie te kunnen toepassen vertrekkend van het berekende model.

Usage

rmse.verschuiving(Verschovenmodel, Boomsoort, Domein)

Arguments

Verschovenmodel

Afgeleid model voor één boomsoort-domeincombinatie (lm-object)

Boomsoort

BMS

Domein

DOMEIN_ID

Value

Dataframe met BMS (boomsoort), DOMEIN_ID, RmseVerschuiving

Examples

library(dplyr)

#Datasets inladen en het basismodel en afgeleid model berekenen
Data <- testdataset()
Datalijst <- initiatie(Data)
Data.basis <- Datalijst[["Basis"]]
Basismodel <- fit.basis(Data.basis)
Data.afgeleid <- Datalijst[["Afgeleid"]]
Afgeleidmodel <- fit.afgeleid(Data.afgeleid, Basismodel)

# De RMSE berekenen voor de verschuiving van het Vlaams model naar een
# afgeleid model
Afgeleidmodel[[1]] %>%
  rowwise() %>%
  do(
    rmse.verschuiving(.$Model, .$BMS, .$DOMEIN_ID)
  ) %>%
  ungroup()
#> # A tibble: 2 × 4
#>   BMS      DOMEIN_ID nBomenModel RmseVerschuiving
#>   <chr>    <chr>           <int>            <dbl>
#> 1 testboom G                  44             1.64
#> 2 testboom H                  18             1.15
#Nota: voor een berekening van de volledige RMSE van een afgeleid model
#moet ook de RMSE van het basismodel in rekening gebracht worden