Genereert testdataset

Deze functie genereert een testdataset die voldoet aan de opgegeven vereisten. De testdataset kan gebruikt worden om het package te testen, ze voldoet als invoer voor functie initiatie()

Usage

testdataset(
  nBomenDomein = c(200, 100, 500, 100, 100, 100, 50, 20, 8),
  BMS = "testboom",
  IDbms = 1,
  minOmtrek = 20,
  maxOmtrek = 239
)

Arguments

nBomenDomein

Vector met aantal gemeten bomen per domein

BMS

Naam van boomsoort

IDbms

Identificatienummer van boomsoort

minOmtrek

Minimale omtrek voor de data

maxOmtrek

Maximale omtrek voor de data

Value

Dataframe met testdata voor 1 boomsoort met velden BMS, IDbms, DOMEIN_ID, BOS_BHI, C13, HOOGTE, Status en ID.