Berekent modelparameters voor opgegeven model

Functie die de modelparameters berekent op basis van een opgegeven basismodel, lokaal model of afgeleid model. Ze berekent de parameters voor het domeinmodel en ingeval van het basismodel ook voor het Vlaams model, en geeft ook de grenzen van het bruikbaar interval.

(Deze functie verwijst naar de interne functies modelparameters.basis(), modelparameters.lokaal() of modelparameters.afgeleid(), afhankelijk van de situatie)

Usage

modelparameters(Basismodel, Data = NULL, Afgeleidmodel = NULL)

Arguments

Basismodel

model per boomsoort (basismodel) of model per boomsoort-domeincombinatie (lokaalmodel)

Data

meetgegevens (enkel nodig voor lokaal model)

Afgeleidmodel

voor de berekening van de modelparameters van het afgeleid model moeten zowel het basismodel als het afgeleid model gegeven worden, dus in dit geval wordt hier het afgeleid model meegegeven. Voor de andere modellen mag dit argument niet toegevoegd worden.

Value

Dataframe met parameters voor domeinmodel (Ad, Bd en Cd) en ingeval van het basismodel de parameters voor Vlaams model (Avl, Bvl en Cvl). Ingeval van een afgeleid model worden de parameters voor het Vlaams model gegeven (Avl, Bvl en Cvl), en een parameter Ad die de verschuiving van het Vlaams model naar het afgeleide domeinmodel weergeeft (dus een extra intercept)