Validatie van een uitbreiding van een model naar hogere omtrekklassen

Functie die toelaat om de opgegeven basismodellen of lokale modellen te valideren ter hoogte van omtrekklassen boven het bruikbaar interval ingeval hiervoor voldoende gegevens beschikbaar zijn. De functie berekent de waarden diff_min, diff_mediaan en diff_max en geeft een overzichtsrapport met de 20 slechtste niet eerder afgekeurde curves (= curves met hoogste absolute waarde voor diff_mediaan), waarvan het de bedoeling is dat de gebruiker ze valideert.

Usage

validatie.uitbreiding(
  Model,
  Dataset,
  AantalDomValidatie = 20,
  GoedgekeurdeUitbreidingen = NULL,
  Bestandsnaam = "Default",
  TypeRapport = "Dynamisch",
  PathWD = getwd()
)

Arguments

Model

Basismodel per boomsoort of lokaal model per boomsoort-domeincombinatie zoals teruggegeven door de functies fit.basis() of fit.lokaal()

Dataset

Dataset zoals teruggegeven door functie initiatie(), op basis waarvan het opgegeven model (basismodel of lokaal model) berekend is.

AantalDomValidatie

Standaard worden de 20 domeinen met de hoogste gemiddelde afwijking (DiffMediaan) geselecteerd om te valideren. (Hierbij worden GoedgekeurdeUitbreidingen niet meer getoond.) AantalDomValidatie laat toe om dit aantal van 20 domeinen aan te passen.

GoedgekeurdeUitbreidingen

Optie om goedgekeurde uitbreidingen niet meer te tonen in het validatierapport zolang er geen extra metingen toegevoegd zijn voor de boomsoort-domeincombinatie (om als gebruiker enkel de te keuren curves in het rapport over te houden). De goedgekeurde en dus te negeren curves moeten opgegeven worden in een dataframe met velden DOMEIN_ID, BMS en nBomenTerugTonen, met nBomenTerugTonen het aantal bomen boven het bruikbaar interval (nExtra) in de boomsoort-domeincombinatie vanaf wanneer de curve terug getoond moet worden. (In dit geval wordt de curve uiteraard enkel terug getoond als ze bij de 20 slechtste curves hoort, met hoogste diff_mediaan.)

Bestandsnaam

Een naam voor het validatierapport (.html-bestand) dat gegenereerd wordt, bestaande uit een string die eindigt op .html

TypeRapport

Default is "Dynamisch", waarbij de figuren in het .html-bestand kunnen worden aangepast (meetgegevens weergeven door muis erover te bewegen (inclusief ID als deze in de dataset meegegeven is), items uit legende wegklikken, grafiek inzoomen,...). Een andere optie is "Statisch", waarbij de figuren vast zijn.

PathWD

Het path van de working directory, dus het path waarin het validatierapport opgeslagen moet worden (default wordt het in de op dat moment actieve working directory opgeslagen).

Value

De functie genereert een validatierapport (.html-bestand) in de working directory met informatie en grafieken van de te controleren modellen. De weergegeven grafieken zijn gerangschikt van slechtste (= mediaan heeft grootste afwijking) naar minder slecht. Idee is om in dit rapport het blauwe deel van de grafiek te valideren.

De functie geeft een dataframe terug met per boomsoort-domeincombinatie de omtrekklasse tot waar het model uitgebreid kan worden (met velden BMS, DOMEIN_ID en MaxOmtrek). Deze tabel kan als argument Uitbreidingen in functie outputIVANHO() toegevoegd worden om deze uitbreidingen mee te nemen voor het eindresultaat. Belangrijk is hierbij dat afgekeurde curves eerst verwijderd zijn door bij initiatie() via Uitzonderingen de variabele min_uitbreiden_model op te trekken en daarna validatie.uitbreiding() opnieuw gerund is met de aangepaste dataset.