Stelt de slechtste curves en afwijkende metingen grafisch voor in een rapport

De functie genereert een validatierapport (.html-bestand) in de working directory (of opgegeven directory) met informatie en grafieken van de te controleren modellen. De afwijkende metingen en curvedelen zijn in rood aangeduid; boven de curve is het probleem ook woordelijk beschreven.

Deze functie rendert het bestand Validatierapport.Rmd, dat afhankelijk van de opgegeven variabele TypeRapport voor elke boomsoort-domeincombinatie als child DomeincurveDynamisch.Rmd of Domeincurve.Rmd toevoegt.

Usage

validatierapport(
  SlechtsteModellen,
  AfwijkendeMetingen,
  Dataset,
  Bestandsnaam = "Validatie.html",
  TypeRapport = c("Dynamisch", "Statisch"),
  Uitbreidingsrapport = FALSE,
  verbose = TRUE,
  PathWD = getwd()
)

Arguments

SlechtsteModellen

Lijst van de slechtste modellen, dit zijn modellen met hoge RMSE, afwijkende vorm (op basis van extremen en buigpunten) en/of modellen met afwijkende metingen. Deze dataframe moet volgende velden bevatten: BMS (boomsoort), DOMEIN_ID en Reden (= tekstuele opsomming van afwijkingen, om weer te geven boven grafiek).

AfwijkendeMetingen

Lijst met afwijkende metingen (dataframe zoals gegenereerd door de functie afwijkendeMetingen()).

Dataset

Dataset met gemeten waarden en geschatte waarde voor domeinmodel en Vlaams model (inclusief RMSE)

Bestandsnaam

Een naam voor het validatierapport (.html-bestand) dat gegenereerd wordt, bestaande uit een string die eindigt op .html

TypeRapport

Default is "Dynamisch", waarbij de figuren in het .html-bestand kunnen worden aangepast (meetgegevens weergeven door muis erover te bewegen (inclusief ID als deze in de dataset meegegeven is), items uit legende wegklikken, grafiek inzoomen,...). Een andere optie is "Statisch", waarbij de figuren vast zijn.

Uitbreidingsrapport

Gaat het over een validatierapport voor validatie.uitbreiding()? Default is FALSE, wat betekent dat het validatierapport de standaard layout heeft voor de validatierapporten, met afwijkingen in rood en standaard informatie die getoond wordt. In het validatierapport voor validatie.uitbreiding() is de uitbreiding die gevalideerd moet worden, in blauw weergegeven. Ook wijkt de weergegeven informatie iets af van de standaard.

verbose

Dit geeft de toestand van het systeem aan en zorgt ervoor dat boodschappen niet onnodig gegeven worden. (Defaultwaarde behouden.)

PathWD

Het path van de working directory, dus het path waarin het validatierapport opgeslagen moet worden (default wordt het in de op dat moment actieve working directory opgeslagen).

Value

De functie genereert in de working directory (of opgegeven directory) een rapport (.html) met de te controleren modellen. Hierin wordt per model (boomsoort-domeincombinatie) de volgende algemene informatie vermeld: boomsoort, domein (en ID), aantal metingen, RMSE, bruikbaar interval en de mogelijke problemen die bij het model optreden.

Daaronder wordt telkens grafisch volgende info weergegeven:

een puntenwolk die de metingen voorstelt (geen individuele metingen, maar een jitter)

curve van het Vlaams model (als beschikbaar, dus niet voor het lokaal model)

curve van het domeinmodel

grenzen van het bruikbaar interval (curves eindigen bij de klassenmiddens die overeenkomen met deze grenzen)

afwijkende metingen: in rood (andere metingen in zwart)

afwijkende deel van een curve in rood (rest van curve in zwart)

Bij de keuze voor een uitbreidingsrapport (argument Uitbreidingsrapport = TRUE) wijkt het rapport af:

algemene informatie: de RMSE is vervangen door de variabelen DiffMediaan, DiffMin en DiffMax

kleur van grafiek: geen afwijkingen in rood, wel is de uitbreiding weergegeven in blauw (zowel jitter van metingen als curve)